Revert "Rename rowGroupMatches to stripeMatches (#2398)" (#2436)
Summary:
This reverts commit bf57c9a95facf012e0dc288363ffa3ae8e94cf32.

Stripe is a specific structure in DWRF and ORC, and each stripe has
multiple RowGroups, or called "strides", which are 10000 continuous
rows. Parquet doesn't have Stripe concept, but just RowGroups. A
Parquet file can also be considerred as a file with only 1 Stripe.
Similar to filterRowGroups(), rowGroupMatches() is also meant for
eliminating specific RowGroups in either DWRF or Parquet, NOT Stripes.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2436

Reviewed By: kgpai

Differential Revision: D39209118

Pulled By: Yuhta

fbshipit-source-id: 5ca42a5f9b3c50d8b6f6b6e34f03914ffb91994e