Fix lock order issue detected by tsan in LocalPartition (#2535)
Summary:
There is a lock order issue detected by tsan in LocalPartitionTest.maxBufferSizeGather
We fulfill the promises in LocalExchangeMemoryManager under the
LocalPartitionExchange's lock and the continuation of the promise is executed inline
which will grab the task lock. This is in reverse to the lock ordering on task startup which
grab LocalPartitionExchange's lock to add producer while holding the task lock.

Verified the fix by running LocalPartitionTest.maxBufferSizeGather under tsan mode on
meta internal devserver for 100 iterations.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2535

Reviewed By: Yuhta, mbasmanova

Differential Revision: D39507965

Pulled By: xiaoxmeng

fbshipit-source-id: 9e18551f5232b20e8803e0001db54e72de0bd621