Parallelize Join Hash Table Build (#2406)
Summary:
The majority of wall time in some TPC-H queries is taken by single threaded hash join build.
We parallelize hash join build as follows:

Compute a partition for each row to insert. The partition is a range
of hash table tag groups. The partition of each row is kept in a
RowPartitions object in RowContainer.

Build the hash table one thread per partition. Each thread reads all
the partition bytes and picks the rows for which the partition matches
the thread. This is fast because all the partition numbers are
contiguous and only the rows that ar to be inserted are touched by
each thread.

If a linear insert into the last tag group of a partition would not
fit and spill into the first group of the next partition, the insert
is recorded as an overflow and the overflows are inserted on a single
thread after the parallel portion. In this way each partition is
guaranteed a single writer.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2406

Reviewed By: xiaoxmeng

Differential Revision: D39111652

Pulled By: oerling

fbshipit-source-id: 5a03fb103318848e8d90adc476923688da76c721