Fix position independent code setting (#2554)
Summary:
This patch fixes the name of option, which is used to set default value for `-fPIC` when building velox code, it should be prefixed with `CMAKE_` to work properly.

Also, since it is a boolean setting, it should be one of the following:
* 1/0
* TRUE/FALSE
* ON/OFF

Changed to the upper-cased `TRUE` form so that this setting evaluates to the boolean-true in the cmake execution context.

This change does not affect the CI pipelines, because these already seem to pass the `-DCMAKE_POSITION_INDEPENDENT_CODE=ON` manually to the build scripts.

Signed-off-by: Pavel Solodovnikov <pavel.al.solodovnikov@gmail.com>

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2554

Reviewed By: pedroerp

Differential Revision: D39662712

Pulled By: kgpai

fbshipit-source-id: 975af144428a54a62782c8e13f65feb959d70989