Fix FlatVector<bool>::ensureWritable (#2601)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2601

FlatVector::ensureWritable allocates a new `values_` buffer and copies old
values to it when the old buffer is smaller than the required new size. For
Boolean vector, it used to allocate a buffer of the size `newSize / 8` but
then copy-paste up to `newSize` bytes into it. This diff fixes this bug.

Reviewed By: mbasmanova

Differential Revision: D39675804

fbshipit-source-id: 8cabded3070db7c6a081a7fe0b37794350ef1ddf