Add Expr::toSql() method (#2613)
Summary:
Add Expr::toSql() method to get SQL string for the expression.

When an error happens during expression evaluation, we save the input vector to
a file and include the file path along with the output of Expr::toString() in
the exception message. The output of Expr::toString() is not exactly valid SQL,
hence, needs to be tweaked to reproduce the error. It would be more convenient
to also write SQL for the failed expression to another file and include that
file path in the exception message. Then, reproducing the error would require
just reading the vector from one file, reading the SQL from another, then
evaluating the expression. No more manual editing of the expression string.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2613

Reviewed By: kagamiori

Differential Revision: D39718245

Pulled By: mbasmanova

fbshipit-source-id: c5246009f9c63703869f73ad7aa7d8fc57827624