Reduce objects in Presto worker
Presto heavily uses RowGroupIndex for pruning. On some internal
meta deployments RowGroupIndexes are cached in CachingStripeMetadataSource.
On these deployments 25% of the objects have root in RowGroupIndex.
Before this change RowGroupIndex, contained ImmutableList of positions
and each position is a boxed integer.
After this change RowGroupIndex contains the primitive integer array.
This eliminates about 15% of the objects on these deployments.