Add micro batching debug logging
Summary:
Micro batching logging to debug oversized stripes. One of the
important factors in controlling the stripe size in dwrf writer is the input
batch size. If we force a vastly oversized batch into the writer, the last line
of defense is the micro batching logic.

This added logging allows us to debug when thing do go wrong. We choose to make
the logging level INFO because verbose level 1 is way too noisy and once per
micro batch logging is not too bad.

Reviewed By: DanielMunozT

Differential Revision: D39838559

fbshipit-source-id: cb248f475aa26c4441de0ab9f09d9ec676085d8e