Fix velox::functions::jsonExtract() support of wildcard select (#2654)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2654

When there is a wildcard select in the json-path argument, the result of
jsonExtract() should be a folly::dynamic that represents an array.
jsonExtract() right now only generates folly::dynamic of array when the
result contains more than one child, but with wildcard select, arrays with
single child are also possible.

Reviewed By: Yuhta

Differential Revision: D39833281

fbshipit-source-id: 97c1f0253f0991b026baf347f68d88eee987885b