Add new RowContainer::extractColumn APIs (#2425)
Summary:
This PR adds 2 variations of the RowContainer::extractColumn API :

i) Value window functions like nth_value need to copy from a RowContainer into a result vector. The rows being copied need not be contiguous. So we are adding a new API similar to extractColumns where we specify in an offsets buffer
the indexes of the rows to be copied.

ii) Another version of the extractColumn APIs used in window functions requires to copy contiguous rows but at a non-0 offset in the resultVector.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2425

Reviewed By: Yuhta

Differential Revision: D39861261

Pulled By: mbasmanova

fbshipit-source-id: a5fc39a220a66955507bff7b810e0c564e23e0f3