Do not throw exception for Parquet corrupted statistics
TupleDomainParquetPredicate.getDomain() is used for Parquet RowGroup
elimination. It could throw ParquetCorruptionException if the Parquet
file column statistics is corrupted and cause the query to fail. This
commit changes the exception to a PrestoWarning with HiveWarningCode
HIVE_FILE_STATISTCIS_CORRUPTION for such case, and returns the full
Domain so that the RowGroup won't be eliminated.