[native] Fix the intermediate type of approx_percentile
Before this change, intermediate aggregation node for
`approx_percentile` in worker only sees a type signature of `VARBINARY
-> VARBINARY`, thus not be able to figure out the actual value type and
cannot perform the merge properly when the value type is not `DOUBLE`.

This fix changes the intermediate result type to `ROW` so that we can
get the value type from it and instantiate accumulator with proper type.

Fix https://github.com/facebookincubator/velox/issues/2430