Fix DecodedVector::dictionaryWrapping() (#2678)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2678

DecodedVector::dictionaryWrapping() was returning nulls_ in all cases when
there was more than one wrapper.
After that the nulls vector was used in forming a new wrapper areound the
base vector which lead to nulls being incorrectly applied and wrong results.
In case when all the wrappers we have are without nulls
(hasExtraNulls_ is false) we should return no nulls byffer (nullptr).
This fix does exactly that.

Reviewed By: mbasmanova

Differential Revision: D39918799

fbshipit-source-id: bf2269b231ca742700c7bf21138f07e13c5302b6