Append selected authorized user to identity
PR (#18223) creates selectAuthorizedIdentity api and invoke it from DispatchManager.
This PR aims to utilize the selected identity in all the subsequent permission checks.

In order to minimize the potential impact to current execution flow, we will add the
selected username into the existing identity, and apply this selected username
for permission checks only.