Fix test generating umbounded float filters (#2623)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2623

testValues and testDouble/testFloat of a float Filter with no bounds
produce different results. We check that a float filter has at least
one bound at construction since range filter with no bounds does not
make sense.

We fix FilterGenerator to generate a not null instead of a
float/double range if the bounds of the range would both be #nan.

Reviewed By: Yuhta

Differential Revision: D37327248

fbshipit-source-id: b469c6003b5ebca232910a94de7b7957ac7712ac