Add support for installing and build dependencies, install folly using FetchContent (#2413)
Summary:
This PR creates a macro `resolve_dependency` to allow us to download/build/install dependencies using `FetchContent`  and migrates the `Folly` dependency to use that macro.

It also updates all the scripts and CI to not require downloading and builiding `Folly` as a system dependency and to be done as part of the build of `Velox` itself.

The new `resolve_dependency` macro  will allow us to find the dependency via the usage of `find_package` or use a custom `build_dependency(DEPENDENCY_NAME)` macro to download and build the third party dependency if not found. We can also define whether we want to force `find_package` or download and build via the `${DEPENDENCY_NAME}_SOURCE` variable.
Usage:
`resolve_dependency(DEPENDENCY_NAME [REQUIRED_VERSION <required_version>] )`

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2413

Reviewed By: Yuhta

Differential Revision: D39853108

Pulled By: Yuhta

fbshipit-source-id: a403b2316af960b2a96729766e51af2d8e192f23