Rewrite IF(predicate, AGG(x)) to aggregation with mask
IF(predicate, AGG(x)) generates a plan which does aggregation on the
whole dataset then project the aggregation results with a IF expression.
However, this is not efficient, as we calculate the the aggregation for
the whole dataset, and discard part of it.
To avoid this inefficiency, we added this optimization which include the
predicate to the mask of the aggregation.