Automatically update docker images (#2608)
Summary:
This draft adds a GH workflow to automatically build and publish the images used for CI (and the dev image). You can see the result in my [fork](https://github.com/assignUser/velox/pkgs/container/velox-dev).  This workflow is fully automated and does not require credentials (it uses the `GITHUB_TOKEN` generated for the workflow). For custom images/special cases anyone with write access to the velox repo can push a new image version/tag (see [here](https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry#pushing-container-images)).

The workflow is triggered when it, the dockerfiles or one of the relevant scripts changes, both in PRs and on push to main. The images are only update on push to main in facebookincubator/velox repo (to avoid noisy error-notification-emails on forks).

I choose to use different tags instead of different names with version tags as old images will no longer be relevant and this way the "packages" section of the sidebar is not flooded with different packages (that are irrelevant to users/most contributors).

I had to update the check dockerfile to 22.04 as 21.04 is eol.

I did not add the `prestocpp/velox-circleci:sagarm-20220131` image used in `doc-gen` as I could not find a relevant script/dockerfile.

kgpai raulcd

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2608

Reviewed By: xiaoxmeng

Differential Revision: D40104972

Pulled By: xiaoxmeng

fbshipit-source-id: 5384e39b8557f7d9de6ee87c68ae9a9edd684c54