Add support for X-Presto-Prefix-Url to Allow Proxying
If this new header is present in the request, Presto will now read its contents, validate it and then prepend it into all uri fields in the QueryResults. A Reverse Proxy can be configured to auto-populate this header field with its own URL and query params. When the client receives the QueryResults, it will issue any subsequent requests to the newly-constructed URL, and hence go through the RevProxy.