Add unit test of wildcard selection in json_extract_scalar() (#2755)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2755

When there is a wildcard in the json path, Presto's behavior is to always
extract an array of selected items, and hence json_extract_scalar() always
return NULL in this situation. On the other hand, the current
json_extract_scalar() returns the selected item itself if there is exactly one
selected. There are some internal dependencies on the current behavior.
So we add this unit test to ensure we don't break the internal dependency
before we fix the function to follow Presto.

Reviewed By: Yuhta

Differential Revision: D40121438

fbshipit-source-id: 98ce84ccc2fb6b32918d81da92c33eaed864ccd9