Fix loading of dictionary wrapped lazy vectors (#2781)
Summary:
Currently, when attempting to load a dictionary wrapped lazy vector,
the resultant dictionary vector can have invalid internal state for
the following two cases:

1. Lazy vector that is wrapped in multiple layers of dictionary
encoding such that the rows that it needs to be loaded for all
end up pointing to nulls. This results in a zero sized base vector
wrapped in a dictionary layer where the rows that were not selected
to be loaded have default values of their index(set to 0) and null
value(set to not null). This violates the invariant that the index
should point to a valid position in the base vector.

2. Lazy vector that is wrapped in a dictionary with no nulls buffer
and being loaded for zero selected rows. This results in a zero
sized base vector wrapped in a dictionary layer of non-zero size
with indices and nulls buffer initialized to default values and
therefore ends up violating the same invariant as above.

This patch fixes both of the above cases.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2781

Test Plan: Added unit tests

Reviewed By: mbasmanova

Differential Revision: D40192874

Pulled By: mbasmanova

fbshipit-source-id: 98239b6f439026a6f29ed9a9f10b51955f4e9d02