Fix cases of filter-only column scans A filtered column may or may not (#2807)
Summary:
Fix cases of filter-only column scan with is (not) null

A filtered column may or may not produce a value. This is controlled
by the ExtractValue template parameter to visitors. If this is
dwio::common::DropValues, a buffer for retrieving the value is not
allocated.

When common parts of the reader were moved to dwio::common, some of
the DropValues mentions in dwio::dwrf were not renamed. Also, there
was a forward declaration of DropValues in dwio::dwrf, causing the unrenamed locations  to
continue to compile.

In Parquet, when reading multipage RowSets, numValuesBeforePage needs
to be the number of passing rows of the RowSet from previous pages,
instead of the number of values, as no values are being collected.

Adds coverage for filter only columns to E2EFilterTest. One in 5
filtered columns will not be projected out. Accordingly, the column
will appear as length 0 in the result. Deals with these in ownership
checking and result comparison.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2807

Reviewed By: tanjialiang

Differential Revision: D40255911

Pulled By: oerling

fbshipit-source-id: c5a7d329144b240298c8ebf6441b6a409f1c8bf0