Fix dedictionarizing a Parquet string (#2829)
Summary:
A dictionary encoded Parquet string column is returned as a string dictionary
vector. If the encoding is switched in mid-scan to direct, values that have
been accumulated as dictionary indices need to be flattened out so that the
scan can continue adding values.

Adds a test case for the scenario. Dedictionarizing can only occur in multipage
reads at the start of a non-first page. In this situation, nulls for the
dictionary vector are not in the vector but in the null concatenation that
collects a multipage null bitmap. Leave the concatenation in place but put the
nulls back into the reader before calling dedictionarize.

Fixes https://github.com/facebookincubator/velox/issues/2808

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2829

Reviewed By: mbasmanova

Differential Revision: D40333035

Pulled By: oerling

fbshipit-source-id: f8507695a508f6d2effa5ac9549b2ed92314fb0b