Add ability to read/write LazyVector via VectorSaver utility (#2893)
Summary:
This patch adds the ability to serialize a lazy vector via the
VectorSaver utility. The lazy vector saved has the header
(size, type) and the loaded vector (if it was loaded at the
time of serialization). The idea here is to recreate the state
of the loaded vector when expression evaluation errored out
(which initiated its serialization). This would help in quickly
reproducing the error as the lazy vector would end up in the
same state if the same steps are executed that resulted in an
error in the first place. However this also precludes it from
being used in any other context like trying to test a fix.

TestPlan:
Added unit test

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2893

Reviewed By: mbasmanova

Differential Revision: D40526320

Pulled By: bikramSingh91

fbshipit-source-id: cffeb0b4d69c1390a16cabd802c152c0598b0a52