Fix handling of dictionary-added nulls in CastExpr (#2881)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2881

CastExpr peels off input encoding, processes the base vector at non-null rows
(i.e., `nonNullRows`), and wraps the local result again with the peeled wrapping.
Before this fix, it uses a SelectivityVector of all rows (i.e., `rows`) when wrapping
the result. This caused an issue because the peeled wrapping is of the length
`nonNullRows.end()` that could be smaller than `rows.end()`, which is not
allowed by DictionaryVector's constructor. This diff fixes this bug.

Reviewed By: mbasmanova

Differential Revision: D40500476

fbshipit-source-id: 20b74840f224696f8050861e66b6e9e5e72094ca