Fix and make Expr::addNulls() a static function for code reuse (#2896)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2896

Expr::addNulls() provides a common logic for adding nulls to a vector and handles
all kinds of cases of input. It is useful when we evaluate an expression on only non-
null rows and want to add remaining nulls to the result. This logic is needed in
CastExpr too, so we make it a static function to make the code reuse possible.

Reviewed By: mbasmanova

Differential Revision: D40530240

fbshipit-source-id: 6c1053a15fdf2916ee90712c129f66febd2a404d