Fix weighted fair scheduling in disagg coordinator
In disagg coordinator, each coorinator is considering it's own running queries to decide which
resource group should run the next query in weighted fair scheduling. This lead to an issue when
with n coordinators each allowing single query to run for the resource group which has lower weight
and lead to n slots been used by the resource group even if it's weight will allow it to run lower than
n queries.

With the fix, we consider global running query count for the given resource group, which helps decide the
correct number of queries we should run for a lower weight resource group and allow more slots be given to
high priority resource group as per weighted fair scheduling.