Fix cast-to-json-in-try caused by peeling of encodings of array or map elements (#2761)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2761

Casting arrays or maps to JSON processes their element vectors
after peeling of element encodings. When the cast is wrapped in
a Try expression, this peeling causes a mismatch between the
recorded error rows and the actual rows of the prolematic elements.
This diff fixes this mismatch by wrapping the errors vector with the
peeled encoding after processing.

Reviewed By: mbasmanova

Differential Revision: D40131312

fbshipit-source-id: 9e202b738c287ecbb804df847bc20684fe67f3ef