Fix cast-to-json-in-try caused by mismatch between element rows and top-level rows (#2760)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2760

The evaluation of cast of array or map to JSON process their element
vectors directly. Exceptions are kept in EvalCtx::errors_ at the same
rows as the elements that cause errors if exist. If the cast is wrapped
in a try expression, these exceptions are then turned into nulls at the
same rows in the result vector. This lead to a mismatch because
exceptions are kept at element rows whereas values in the result vector
are at top-level rows. This diff transform EvalCtx::errors_ to map
exceptions back to top-level rows for cast to JSON.

Reviewed By: mbasmanova

Differential Revision: D40131130

fbshipit-source-id: 9da45cbcd475edb57611f047b347eaf6545cb4e0