Add ExpressionRunner for failure reproduction (#2876)
Summary:
Expression runner is used to run custom expressions with custom input data.
One of the use case scenarios is when ExpressionFuzzer run fails, it preserves
the failed expression and its input data to disk. This expression runner can then
be used to reproduce the failure for debugging purposes.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2876

Reviewed By: mbasmanova

Differential Revision: D40516702

Pulled By: tanjialiang

fbshipit-source-id: e90e254fab4002df447d04fab69175260e0d2de1