Fix CI for next snapshot iteration PRs
The new release process for presto creates pull requests
that have an incremented snapshot which hasn't
been built by anything yet. There are tests that are building
test coverage but skipping presto server. This creates a problem
where they are requesting dependencies that aren't built yet.

We don't need test-coverage in these tests, so it's best if
we just skip building those modules.