Fix ConstantExpr::toSql for negative numbers, dates and nulls (#2949)
Summary:
Negative numbers and date literals need to be wrapped in single quotes,
otherwise '-' characters are interpreted as negation or minus operators.

Also, null literals need to specify the type. Otherwise, these are parsed as
nulls of UNKNOWN type.

Enhanced ExprTest::constantToSql to evaluate the SQL and compare the
results with evaluating the original expression.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2949

Reviewed By: xiaoxmeng

Differential Revision: D40666445

Pulled By: mbasmanova

fbshipit-source-id: b5b7721ada740ba5a5c028c81912f835ba18d6c5