Build Protobuf with -Wno-invalid-noreturn (#2976)
Summary:
Velox builds with -Werror and -Winvalid-noreturn.  When Protobuf builds it inherits these flags, but at the current release fails this check.  Adding -Wno-invalid-noreturn to skip this check when building Protobuf.

Note that this has been fixed in Protobuf's source, but never made it into a release
https://github.com/protocolbuffers/protobuf/commit/1e9d43fe711b46bcb9bb82849d495f137c9d1ef8

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2976

Reviewed By: pedroerp

Differential Revision: D40735923

Pulled By: kevinwilfong

fbshipit-source-id: 7c68a3fcf23ca196c0e17fc32dab3595a3912074