Map() can only take fast path if offsets are the same
Summary:
Map() function can only take the fast path if there's no indirection
for both keys and values vectors, but also if the offsets are the same (to
ensure the order of elements are the same). In case the sizes are different the
function will already inadvertently throw.

Reviewed By: mbasmanova

Differential Revision:
D40708097

LaMa Project: L1134559

fbshipit-source-id: 04ca325d9befdb6091c5ff44764b98dd277d553d