Fix Protobuf build in dwio (#2979)
Summary:
With the recent change to build Protobuf locally, the dwio/dwrf/proto build has started failing in linux-benchmarks-basic-dedicated.

I think this is due to the fact it was still using the protobuf_generate_cpp calls from FindProtobuf.  I suspect this was working in other cases because it was either finding similar Protobuf releases already installed on the machine or using the one built locally because no other was available.

In the case of linux-benchmarks-basic-dedicated it looks like it was finding an older version of Protobuf on the box that was incompatible with the version we built locally.

I rewrote the CMake file to generate the dwio Protobuf code to use the variables exported by the logic to locally build Protobuf and that seems to have fixed it so that it consistently uses the locally built version (both for the compiler and the linked libraries).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2979

Reviewed By: pedroerp

Differential Revision: D40768683

Pulled By: kevinwilfong

fbshipit-source-id: e45cebd2f68939faa5ae5f138d4a08bbfb178586