Fix nightly fuzzer run timeouts (#3006)
Summary:
The nightly fuzzer run has been timing out with "Too long with no output (exceeded 20m0s): context deadline exceeded"

The problem is that when we run "make fuzzertest", internally it runs "make debug".  Even though we just ran "make debug" in the previous step, it partially rebuilds.   Between this and the 10 minute fuzzer tests it exceeds the 20 minute limit.  Note that since we pipe all output to a file, the 20 minute timer starts right away.

I changed the config for this step to execute the velox_expression_fuzzer_test directly rather than calling "make fuzzertest" since this avoids the build.  I could have increased the time limit, but rebuilding seemed like a waste anyway.

I also used tee to pipe the output to stdout while also writing it to a file (if the fuzzer test succeeds the output gets truncated in the CircleCI logs, but it's helpful to have it readily available if something goes wrong, and this allows us to watch the output at the beginning in realtime (e.g. so I could immediately see that it was rebuilding).  (I moved the message that we're writing everything to an artifact to the top so it doesn't get cut off).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3006

Test Plan:
Here is an example of a successful run with these changes
https://app.circleci.com/pipelines/github/facebookincubator/velox/15571/workflows/ccfed344-ddde-4a64-9947-7c68df1129a6/jobs/95261

(I had modified the circleci jobs that run on PRs to run the nightly fuzzer test so I could test it)

Reviewed By: kgpai

Differential Revision: D40813911

Pulled By: kevinwilfong

fbshipit-source-id: dd4b5f8e26e6b3a20f60b34a7705bb329078ba0f