Update release notes script to specify a version
The release notes script needs to be able to accept an explicit
version so that the release oncall can start making the release
notes immediately without having to wait for the state of master
to be updated. Especially now that the cut process creates a pull
request to update the version of master, instead of committing
stright to master.