Generate not null and unique map keys on VectorFuzzer (#3010)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3010

VectorFuzzer now provides a normalizeMapKeys flags. If true, generated
maps are ensured to have unique and non-null keys. This works by either
generating not null vectors (or checking if the provided key vectors are
not-nullable).
.
To guarantee keys are unique, each map element is checked and cut short if
duplicate values are found.

Reviewed By: mbasmanova

Differential Revision:
D40787363

LaMa Project: L1134559

fbshipit-source-id: a2ce607e7134b8d85c9f778e51c884190c90df9e