Extend ExpressionRunner to better support debugging workflow (#3013)
Summary:
When debugging fuzzer-found bugs, it is often beneficial to try to simplify the
expression and reduce the set of rows for evaluation. The new --sql option
allows to try different expressions without editing files. The new --num_rows
flag allows to specify maximum number of rows to process.

It is also helpful to evaluate the expression and print out the results. This is
now possible using --mode=common or --mode=simplified.

- Add common and simplified modes to evaluate expression using common and
  simplified paths and print out the results.
- Add --sql flag to allow specifying SQL expressions in the command line. Add
  support for multiple comma-separated expressions in common and simplified
  modes.
- Add --num_rows flag to allow restricting the number of rows to process.
  Default: 10 rows.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3013

Reviewed By: bikramSingh91

Differential Revision: D40825206

Pulled By: mbasmanova

fbshipit-source-id: f9b3025e13d104e7aa8d2a592e8fbb4fb82f1e8b