Refactor Window function tests to simplify the combinations (#3020)
Summary:
Removed the expensive testTwoColumnOverClauses method and replaced it with a
simpler test API that takes a combination of (data, function, overClauses) to
do the important tests. Common sets of data functions, overClauses are moved to
the WindowTestBase for reuse. This new refactoring will make it easier to add
the frameClauses also to the testing.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3020

Reviewed By: xiaoxmeng

Differential Revision: D40879060

Pulled By: mbasmanova

fbshipit-source-id: 7999910237f77c2a8c118dee09771f867480e003