Fix "Output buffers for task not found" in PartitionedOutputBufferManager (#3035)
Summary:
This is the case where the downstream prematurely aborts and calls abortTask of
the upstream. The abortTask delete the task asynchronously from the buffers.
This causes and exception due to VELOX_CHECK in the following call tree
(because the buffers does not have the entry for the taskId any more)

Driver.runInternal
  -> PartitionOutput.getOutput()
  -> Destination.flush() (if (blockedDestination) {} path)
  -> PartitionedOutputBufferManager.enqueue()
  -> PartitionedOutputBufferManager.getBuffer()  <- Exception here

In the case I've seen, the abort would happen between the flush() and getBuffer()

The fix is to delete the code that deletes the buffer for the task. The buffer
for the task will be removed in Task::terminate() as part of the normal Task
lifecycle.

Fixes https://github.com/facebookincubator/velox/issues/3009

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3035

Reviewed By: spershin

Differential Revision: D40878334

Pulled By: mbasmanova

fbshipit-source-id: acdc01231c5f7bedd8bcbb0db75aa0ccd660a78e