PrintPlanWithStatsTest.innerJoinWithTableScan is flaky (#3032)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3032

PrintPlanWithStatsTest.innerJoinWithTableScan executes a
query wich a has join and prints the plan with the resulting
stats.

The way HashJoins are executed, the HashBuild and
HashProbe are executed in separate Drivers, and hence on
separate threads.

This means that the HashProbe may or may not have to wait
for the HashBuild to finish.  E.g. if the thread running the
HashBuild finishes before the thread running the HashProbe
starts, then obviously it won't need to wait.

The HashProbe operator will only have the stat
queuedWallNanos if it is queued, i.e. if it is blocked and has to
wait for the HashBuild.

Therefore, since
PrintPlanWithStatsTest.innerJoinWithTableScan assumes the
queuedWallNanos stat is always present, it's flaky.

To address this, I've extended the PrintPlanWithStatsTest to
allow us to mark lines as optional.  In this case, since we can't
know for sure if the stat will be present, that line is optional,
and the test will look for it, and if it can't find it, happily move
onto the next one.

Reviewed By: Yuhta

Differential Revision: D40868959

fbshipit-source-id: d75ef6b75e73d9fe242a8f22801b3e1fc51ffd2e