Add coalesce and if to Fuzzer (#3070)
Summary:
Add --special_form flag to Fuzzer to allow for specifying a subset of special
forms to include in the expressions. Supported special forms are: AND, OR, IF,
COALESCE.

Make sure --only applies to special forms as well as functions.

By default, enable only AND/OR special forms (as before the change).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3070

Reviewed By: xiaoxmeng, kagamiori

Differential Revision: D40941492

Pulled By: mbasmanova

fbshipit-source-id: 52b0662e0e021543f247cd848c129bec3f684067