Task Listener Test Cleanup (#2964)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2964

# Why?
Two objects `TestTaskListener` and `std::vector<TaskCompletedEvent>` must live together, but we instantiate separately and pass a reference. Why should we do this?

It seemed to be ok, but it got ugly when I needed to introduce two listeners. I had double number of symbols to deal with.

Let's get the ownership clear unless there is a reason.

Reviewed By: xiaoxmeng

Differential Revision: D40721328

fbshipit-source-id: a1eb4c9639641124108eed3d6f8d4cc989cc377c