Add min spill run size config to avoid generating small spill files (#3064)
Summary:
Add query config of min spill run size limit to select partitions for spilling.
The spiller tries to spill a previously spilled partitions if its data size exceeds
this limit, otherwise it spills the partition with most data. If the limit is zero,
then the spiller always spill a previously spilled partition if it has any data.
This is to avoid spill from a partition with a small amount of data which might
result in generating too many small spilled files.
Unit test and document updated.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3064

Reviewed By: oerling

Differential Revision: D40924224

Pulled By: xiaoxmeng

fbshipit-source-id: b2ca9de01da2393822399564e6e8b654d5bb6ff9