Fix all nulls in an IN list (#3050)
Summary:
An in list in Presto may consist of a single null. If this is the case the predicate is always null for any input. A in list with nulls only is OK, a values list with no elements is an error.

Adds a signature with unknown type for the case of an all null in list.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3050

Reviewed By: mbasmanova

Differential Revision: D40895757

Pulled By: oerling

fbshipit-source-id: 9cd155cc8b93f7afb3257c7509ef8c7b85a2c2ce