Fix Bug: copy from view types. (#3081)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3081

Before this diff, copy_from view types into a writer a type
used to utilize the vector copy operation.
There is two problems with that when used with nested complex types:
1. The recursive copy operation performs an append operation to the child vectors, (i.e: elements of array).
However, the writers will have it larger than needed does
since it does exp growth for those and tracks the actual used index. Meaning things will be written with
gaps, furthermore if an add_item was called after the copy_from, it won't be appended to the end but on
the tracked offset causing bugs.

2. The recursive copy operation does not update the sizes in the nested writers.

Reviewed By: kagamiori

Differential Revision: D40732089

fbshipit-source-id: f52c5431a671416d2e819dc29ab31f2546a8953e