Add support for lazy vector generation to expression fuzzer (#2918)
Summary:
This patch leverages the ability to wrap a vector in lazy encoding
that was recently added in the VectorFuzzer and enables it to run
the same input in both the common and simplified evaluation path
but with the common path having some inputs wrapped in lazy
encoding. The columns to be wrapped are randomly chosen with a
probability that can be specified by a gflag
"lazy_vector_generation_ratio".

Additionally
- context about which columns were chosen to be wrapped are also
seriazlized with other data required to repro in case of an error.
- The repro data is now stored in a newly generated folder with a
unique name.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2918

Reviewed By: pedroerp

Differential Revision: D40610884

Pulled By: bikramSingh91

fbshipit-source-id: 1961ea1a749d443c3f9153cc55c3fc6100b31408