Add CPP UDF type
Currently Presto has built-in and non built-in UDFs. For non built-in
UDFs, it has three types, SQL, thrift and grpc. For SQL UDF, it will be
evaluated in coordinator, and sent to local worker for execution. For
thrift and grpc UDFs, they will not be evaluated in coordinator and will
be sent to remote servers instead of local workers for evaluation.

In this commit, I add one new UDF type, CPP. This UDF is only used when
the local worker is CPP worker instead of JAVA worker. This UDF will not
be evaluated in coordinator, and will be sent to local cpp worker for
evaluation.

Co-authored-by: Sourav Pal <souravpal@users.noreply.github.com>