Add consistency checks in map and array vectors (#3078)
Summary:
This adds consistency checks to both map and array vectors that
ensure that the sizes of the offset and size buffers are valid.
Moreover, adds a more through check that only fires in debug
builds which checks whether the values in the offset and size
buffers are valid.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3078

Reviewed By: xiaoxmeng

Differential Revision: D41053187

Pulled By: bikramSingh91

fbshipit-source-id: 6c5de9aa60a5b29b4ec840352406e257ebbd02ae