Make Resource Group operations executed in SingleThreadExecutor
Problem description
Presently, all operations on Resource Group tree-like data
structure are protected by the root group monitor. That
makes all operation are blockable. Essentially, these
operations are the bottleneck for the Coordinator. It limits the
number of queries that could be admitted by Coordinator.
Additionally, those operations are invoked from many different
threads that could be scheduled on different cores. That
hurts the performance of those operations due to the
necessity of running costly cache coherence operations to
invalidate local caches on different hardware cores. That
limits the bandwidth on operations over Resource Groups.

Proposed solution
Rewrite the InternalResourceGroup class making all the
operations on resource group tree to be executed in the
Single Threaded Executor. After moving this logic into
executor, all operation will be executed in a single thread that
opens the opportunity for Lock Elusion.