Fix map_keys and map_values presto function (#3080)
Summary:
Currently both these functions generate a result array of size equal
to the total size of the selectivity vector passed to it. However if
the last valid row in the selectivity vector is smaller than its
total size (that is rows.end() < rows.size()) then this can be
larger than the size of the underlying buffers and can result in an
inconsistent state where the size of the offset or sizes buffer is
insufficient to hold all indices. This patch fixes the
aforementioned bug.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3080

Test Plan:
Added a unit test that would fail without this fix
(but needs https://github.com/facebookincubator/velox/issues/3078 for consistency check)

Reviewed By: mbasmanova

Differential Revision: D40988298

Pulled By: bikramSingh91

fbshipit-source-id: d98694d1c942aca0ad76ef81842171bdfdb11f56