Fix leaked query count reporting in the disagg coordinators codepath
Summary: Currently in disaggregated codepath, all running queries show up as leaked queries. After this change, the number of leaked queries reported by coordinators in the disaggregated codepath is at least eventually consistent: Due to the eventual consistence nature of the design, some amount of false positives of leaked queries are expected since the coordinator has a stale view of the running queries and the memory state across each worker.

Also made the leak check logging be WARN logging and pulled out the check to be every minute instead of it being every second.

Added logging of the leaked bytes in addition to the leaked queries.

Test plan - (Please fill in how you tested your changes)

Added unit tests.

Checked the metrics in a production environment, that the number of leaked queries is now usually one or two both in disaggregated and normal codepaths.