Use VELOX_CHECK instead of C++ assert in reverseUnicode() (#3178)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3178

Failure of `assert` causes SIGABRT and cannot be caught by C++ try-catch blocks.
This is not compatible with fuzzer test because it fails the test immediately instead of
holding the exceptions from the common and simplified paths and comparing them
later. This diff replace the C++ assertions with VELOX_USER_CHECK.

Reviewed By: xiaoxmeng

Differential Revision: D41148068

fbshipit-source-id: e6977297e7ea04077c81b1d4eacc5c1ad1018032