Fix the order of grouping keys in GroupIdNode's output (#3137)
Summary:
The output type of GroupIdNode is constructed by the outputGroupingKeyNames,
which is stored in map and will be sorted by map key. So when the grouping sets
keys are out of order, the output type will be wrong. This PR use int as the
key to fix the above wrong output type issue.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3137

Reviewed By: laithsakka

Differential Revision: D41173203

Pulled By: mbasmanova

fbshipit-source-id: fb0627839dda76c17e3f964480d620c03ce884c8