Avoid footer stats when not reliable
For partition that goes through mutation via commits, the footer stats
may not be reliable to use for optimization. As part of this commit, we
are providing an option to avoid using footer stats.
Avoid footer stats for metalake partition

For metalake partition, the footer stats are not reliable. As part of
this commit, we are avoiding to use footer stats if a particular split
has underwent mutation.