Fuzzer: Add to results verification for order-dependent functions (#3209)
Summary:
For some functions, the result can be transformed to a value that doesn't depend
on the order of inputs. If such transformation exists, it can be specified to
be used for results verification.  For example,

- array_agg(x) -> array_sort(array_agg(x))
- map_agg(x) -> array_sort(map_keys(map_agg(x)))
- map_union(x) -> array_sort(map_keys(map_union(x)))

This additional verification caught a bug in map_union: https://github.com/facebookincubator/velox/issues/3206

Also, add --num_batches command line argument to control the number of input vectors used for testing.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3209

Reviewed By: xiaoxmeng

Differential Revision: D41230197

Pulled By: mbasmanova

fbshipit-source-id: a1fd16aa9a1bdf7adcbb56d7ba0b1754218d3e9e