Refactor HiveTableLayoutHandle
HiveTableLayoutHandle has 2 constructors, one with partition and
one without partition. If callers use the wrong constructor it
will throw.

HiveTableLayoutHandle is modified at multiple places and all
these places would change one or two parameters, but copy the
rest of the other parameters. It is very hard to say, what is
the intentional change what others are modified.

When introducing new parameters, all places change and it is hard.

The main breaking reason for this PR:
When there are classes that derive from hiveTableHandleLayout these
modifications convert them to base class and mess up the behavior.

To fix all of this, the class supports builder pattern for mutation.