Add pybind support (#3031)
Summary:
Pybind11 support is required for velox python bindings. This initial PR adds support for getting pybind. Subsequent PRs will use pybind11 and setup the bindings.
Note: The expectation is that most of the times the python package will not be built and hence its not required to add pybind as a submodule.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3031

Reviewed By: pedroerp

Differential Revision: D40911873

Pulled By: kgpai

fbshipit-source-id: 9de75cd2fbe4d69e599e78c4760fcb99bcbc03b0