Add check to enforce invariant related to lazy vectors (#3214)
Summary:
This adds an explicit check to ensure an unloaded lazy vectors cannot
be wrapped by two different top level vectors, that is
Dictionary1(lazy1) and Dictionary2(lazy1) is not possible but
Dictionary3(Dictionary1(lazy1)) is possible. This is an attempt
to reduce the possibility of a lazy being loaded with potentially
different set of rows.

Limitation:
This does not prevent the case where another reference to the lazy can
potentially call load on it.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3214

Test Plan: Added a unit test.

Reviewed By: mbasmanova

Differential Revision: D41278982

Pulled By: bikramSingh91

fbshipit-source-id: cde8def0bc3980fc5e672f3a24e561b85af019c6