Fix last scheduling cycle run time metric
For disagg coordinator, when cluster restarts and if resource managers take some time to come up, coordinators will end up having a spike
in scheduling cycle run time due to coordinator not able to reach to resource managers. With this change, we avoid having this spike
and return 0 value till coordinator able to reach to resource manager first time.