Ensure slice() does not generate overlapping ranges (#3253)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3253

In cases where the first parameter is a constant, but not the second
or third, slice() was generating overlapping (offset, size) ranges, which ends
up breaking other functions upstream that rely on the fact that array elements
do not overlap (like reverse). To prevent these cases, flatten the constant input.

Reviewed By: mbasmanova

Differential Revision: D41330251

fbshipit-source-id: d062e9c48d22a0b80255eb51b9e290394b87c385