Enhance lazy vector generation in vector fuzzer (#3166)
Summary:
Currently vector fuzzer exposes an API that allows wrapping an input
vector with a lazy layer. If an input vector with multiple
dictionary layers is wrapped in this lazy layer then when it is
loaded, it combines all the dictionary layers into one similar to
what a decoded vector does. This takes away the ExpressionFuzzer's
ability to have lazy inputs with dictionary layers on top.
This change allows the fuzzer to identify this kind of input vector
and only apply the lazy layer on the innermost dictionary layer.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3166

Test Plan: Modified an existing test case

Reviewed By: mbasmanova

Differential Revision: D41146404

Pulled By: bikramSingh91

fbshipit-source-id: 46740e7f3aa2a5e9956e27ced792ce225fea6cac