Window aggregate optimization for unbounded preceding frames (#3200)
Summary:
When doing aggregation for windows where the frameStart remains fixed and the frameEnds are non-decreasing (like unbounded preceding to current row), there is an additional potential to optimize if this condition holds across the consecutive output blocks of the partition. In such a case, the aggregation state from the previous block can be retained, only argument values for the new frame rows in the current block are read (instead of from the beginning) and incremental aggregation is resumed.

Note : New tests are not required. The different code-paths are already tested on account of the new frame testing for aggregates.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3200

Reviewed By: amitkdutta

Differential Revision: D41380471

Pulled By: mbasmanova

fbshipit-source-id: 57fd716f8a4c1da60ad9310737d429d0850a51be