Centralize logic to convert CallExprs to SpecialForms (#3228)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3228

Currently logic related to allowing SpecialForms to be
expressed as CallExprs is spread across two files.

1) TypeResolver has special logic for each special form to
simulate type resolution of them as function calls.
2) ExprCompiler has special logic for each special form to do
the conversion from CallExprs.

(Side note, I'm planning to add a third for function resolution)

Rather than writing logic specific to certain special forms in
files not directly related to those special forms, and
duplicating the knowledge of what special forms exist across
those files, I propose:
1) Creating a single file FunctionCallToSpecialForm that
knows what SpecialForms can be expressed as CallExprs,
and exposes a single interface to do that special handling
for all SpecialForms.
2) Moving the type resolution and the logic to convert to
SpecialForms to the files specifying those SpecialForms.

This makes it easy to share type checking logic between the
SpecialForm constructor, the type resolution and the logic to
convert from a CallExpr.

This means the names and list of SpecialForms only need to
be hard coded in one place.

This consolidates logic for SpecialForms in a few places
directly related to those SpecialForms.

Reviewed By: mbasmanova

Differential Revision: D41282078

fbshipit-source-id: 234611bcc8533f5d27be814c7029445a3ad71b22