Disable spilling for distinct aggregation and pre-grouped aggregation (#3254)
Summary:
Spilling for distinct aggregation has following problem:
The current implementation for distinct aggregation has assumption
that all the groups are kept in the hash table for the entire processing
so we can start yielding result immediately and output the new distinct
groups as new groups inserted in hash table. This assumption is no longer
true with spilling enabled. As we might turn on spilling in the middle of
input processing, there is no way to tell if a new inserted group is new
distinct group or not as the hash table only holds a subset of groups.
The proposed solution is described in [3263](https://github.com/facebookincubator/velox/issues/3263)

Spilling for pre-grouped keys has following problem:
The current pre-grouped key hash aggregation implementation starts to
yield aggregation result once after it processed all the input rows with
the same pre-grouped keys. It works like streaming aggregation and won't
hold much in-memory state. But in rare case that some particular
pre-grouped keys has very large number of rows, it can still trigger spilling.
Once it happens, we can not simply yield the result after process all the rows
with the same pre-grouped keys as the spilling is triggered in the middle of
row stream.
The proposed solution is described in [3264](https://github.com/facebookincubator/velox/issues/3264)

This PR disable the spilling for the above two hash aggregation use cases
and add unit tests to cover.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3254

Reviewed By: mbasmanova

Differential Revision: D41332015

Pulled By: xiaoxmeng

fbshipit-source-id: 7975f775c918ca585cf9774e86c298ee18415fb9