Fix wrong results in MetadataQueryOptimizer with subfield filters
Previously, queries with aggregations on partition columns and equality
filters on row subfields that had optimize_metadata_queries could return wrong
results.

The MetadataQueryOptimizer rewrite should not take effect if there is a
filter on a non-partition column that was pushed inside the scan.  The
way we check for these filters is by looking at the predicate and
remainingPredicate in ConnectorTableLayout

When converting a HiveTableLayoutHandle to ConnectorTableLayout, equality
filters on row subfields were being left out. As a result, we were incorrectly
applying the MetadataQueryOptimizer rewrite when there was an equality
filter on a row subfield.

An example affected query shape is:
SELECT max(partition_column)
FROM my_table
WHERE struct.subfield IS NOT NULL