Adds Parquet list reader for scalars (#3334)
Summary:
Adds interpretation of repetition and definition levels for nested data to PageReader. Levels are read for the leafmost column of a list and stored in the PageReader for the column chunk on first use.

the repeated reader is inherited from dwio/common, formerley SelectiveRepeatedColumnReader in dwio/dwrf/reader/. The Parquet nested types read like the ORC ones once the repdefs are read for the column and converted to lengths and nulls at the different levels. This is done on first use of a non-scalar reader under the top struct reader.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3334

Reviewed By: Yuhta

Differential Revision: D41475318

Pulled By: oerling

fbshipit-source-id: eccf6587d1b2d7adc42eb7feec151186e50980e1