Enhance fuzzer to re-use sub-expressions and columns (#3352)
Summary:
After this change the expression fuzzer can generated sub-expressions
that re-use some already generated columns. Before picking an input
column to an expression it will check if there are existing columns
with the same type and randomly pick whether to re-use one of them
or create a new one.

Similar functionality is employed to re-use sub-expressions.
However, there is a current limitation to this that only expressions
with no subexpressions are shared.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3352

Test Plan:
Ran fuzzer tests locally and verified manually that columns and
sub-expressions are being re-used.

Reviewed By: kagamiori

Differential Revision: D41561631

Pulled By: bikramSingh91

fbshipit-source-id: 2f21cd1ab16f0402edd05ad5a66e702ed141c6be