Fix template for If expression used in the fuzzer (#3376)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3376

Currently the template for If expression contains one extra input
parameter for each case, the first case (using 3 but expected 2)
works fine because it assumes the extra param is an else clause,
but for the second case (using 4 but expected 3), it expects a
boolean as return type for the third param but gets 'T' and throws
a USER error which both the common and simplified path throw.
Hence it gets ignored as a user error and the fuzzer never crashes.
This patch fixes this bug.

Reviewed By: kgpai

Differential Revision: D41600721

fbshipit-source-id: c0dc1a01eef5c0cd261fc73d041bebb7d1908ed9