Disable S3 Select pushdown when the query does not filter data
S3 Select pushdown improves queries performance by reducing the data
processed on the client side and it is effective when the query pushed
down to Select is filtering out a significant portion of the data.
This commit disables Select pushdown when the query does not have
a predicate or projection. In these cases using GET is both cheaper
and faster.