Fix hash table hang with large number of tombstone slots (#3337)
Summary:
In spilling stress test with ~256KB memory limit when replaying meta production
queries in staging cluster, we run into hanging issue and it turns out the the
hash aggregation operator is stuck in hash probe loop. RCA is that the spilling
execution path triggers hash table entry deletion which leaves a large number
of tombstone slots in the table and a tombstone slot can be used as free slot if
the associated table slot group has no empty slot. The reason is the hash table
is designed to leverage the tombstone slot to handle hash table collision. When
we lookup hash table, if a table group has no target table slot and no empty slot,
then we have to continue search the table as there might be potential match in
the rest of table slots. Given that, when we delete a table slot, we have to mark it
as tombstone slot instead of empty slot. A tombstone slot be used to store new
entry only after the lookup eventually hits a group which has empty slot. And the
first encountered tombstone slot during the lookup process will be used to store the
new entry. So if all the table slots which are either occupied or tombstoned, then
we can't have any free slot to store the new table entry. So we need to treat tombstone
as an occupied slot when we decide to trigger rehash or not. This will also help
improve lookup performance if there is a non-trivial amount of tombstone slots in the
table.

This PR fixes the issue by counting the tombstone slots on deletion/insertion and treat
tombstone slots as an occupied slot to trigger rehash in HashTable::checkSize(). A unit
test is added to reproduce the issue and verify the fix. Will verify on staging cluster with
spilling stress workloads.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3337

Reviewed By: mbasmanova

Differential Revision: D41477801

Pulled By: xiaoxmeng

fbshipit-source-id: a3bbf40ffbd324c5b06c9013775ff19e1b43374d