Fix an invalid check in hash table resize (#3496)
Summary:
There is a case the hash table check size see the table size
matches the number of distinct entries that it stores (see unit test
added in this PR). In the last check size, new distinct + existing
distinct == table size, then the new table size doesn't change
(next power of two of new distinct + existing distinct == table size).
This triggers a recent added check failure.

This PR fixed this issue by ensuring we always bump to the next
power of two even if the new size (new distinct + existing distinct)
is power of two. With the fix, we keeps the existing strict check
condition. We add to expose the size of the table. Unit test is added
to reproduce the issue and verify the fix.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3496

Reviewed By: spershin, oerling

Differential Revision: D41979958

Pulled By: xiaoxmeng

fbshipit-source-id: 7dcc5364e5059bba2ff8d29ce0ba9f55aae3aa46