Adding Task::spillDir_. (#3494)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3494

Adding Task::spillDir_ which will hold the unique spilling
path for the task. This path would be given to the Task by the
application that uses Velox.
Using Task::spillDir_ to construct the complete spilling path for
the operator.
Currently in Presto Native the Task::spillDir_ would be empty until
the follow up change is landed which set it during the Task creation.

Reviewed By: xiaoxmeng

Differential Revision: D41895734

fbshipit-source-id: 0217315d0160fc36f6d00e4446dd88b0672254b1