Do not throw from PartitionedOutputBufferManager::getData (#3502)
Summary:
A task being executed by multiple drivers, and if one of the drivers catches an
exception, can terminate the task. This termination will remove any
PartitionedOutputBuffer assigned to that taskId. In a parallel thread, the task
manager may try to access this buffer leading to "Output buffers not found"
exception. This exception is valid, but instead of the exception encountered by
the Driver trickling up to the user, the "Output buffers not found" exception
may sometimes get thrown to the user.

Instead of throwing exception, the getData() will return
a bool indicating whether a matching buffer for a given task id was found or not.

See https://github.com/facebookincubator/velox/issues/3009 for more context.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3502

Reviewed By: Yuhta

Differential Revision: D42102546

Pulled By: mbasmanova

fbshipit-source-id: 72fab8dd61d8217523ddb55f8d199746a003c027