Support resolving SpecialForms (#3229)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3229

Velox supports expressing some SpecialForms as normal
function calls (CallExprs).

However, Velox does not support resolving the "function
names" like normal functions, users have to add special
handling for these, which takes away some of the magic and
puts a burden on users.

To help with this, I've added a new function
resolveFunctionOrCallableSpecialForm
that can resolve functions are SpecialForms that can be
written like function calls.  Here we're using a new function
to preserve backwards compatibility of resolveFunction.

This allows users the freedom to not distinguish between
function calls that are really functions and function calls that
are special forms.

The one exception is CAST.  The return type for a CAST
cannot be determined from its input arguments.  The type
being casted to is a special argument that is not part of the
function call.  For this reason users still need to have special
handling for this case.  With this change they will get an
explicit error message that CAST does not support type
resolution, instead of a nullptr as they would with
resolveFunction.

Reviewed By: kKPulla

Differential Revision: D41283895

fbshipit-source-id: 2ef2d8ed0e54c651278d8185e1e5ae1d37188b5e