fuzzMap calls in VectorFuzzer yield inconsistent results across platforms (#3561)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3561

The order arguments are evaluated in a function call in C++ is not specified.

This leads to issues when we call functions that invoke our RNG in the VectorFuzzer in two arguments, like these fuzzMap calls where we call one such function for the keys and one for the values.

This came up when trying to repro
https://github.com/facebookincubator/velox/issues/3419

In the CircleCI environment the values were generated before the keys when calling fuzzMap, while in my local environment the keys were generated before the values.  I was getting different results and unable to reproduce the issue.

Constructing the arguments outside the function call allows us to control the order in which their invoked.

Reviewed By: bikramSingh91, kagamiori

Differential Revision: D42178884

fbshipit-source-id: 15bc02b9b275bc0914c1538754c7260a01406090