Upload dedicated target benchmark results to Conbench (#3452)
Summary:
This PR adds an `--conbench_upload_run_id ` flag to `scripts/benchmark-runner.py` that makes the script upload benchmark results to Conbench after running. It also re-enables this uploading behavior for the target benchmarks that are run on the dedicated hardware.

Under the hood, this `--conbench_upload_run_id ` code uses a package called `benchadapt`, which is a Conbench client intended for situations like this (i.e. adapting Folly results to Conbench). It has a special class for dealing with Folly benchmark results, so we can rip a lot of the old boilerplate code out. This PR completely removes the `veloxbench` package from this repo, and only adds a tiny portion of it to `scripts/benchmark-runner.py`.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3452

Reviewed By: kgpai

Differential Revision: D42197199

Pulled By: pedroerp

fbshipit-source-id: b612e126a326fd77cc3703b7ef5cb852743f4d59