Allow function signature templates to be used at top-level expression in fuzzer (#3540)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3540

The signatures of some functions contain type variables such as "T". When generating
a random expression, fuzzer currently choose the top-level function only from function
signatures that do not contain type variable. This is a limitation because sometimes we
would like to test only a function that have type variables in its signature.

This diff allows choosing function signatures that contain type variables as the top-level
function. It also add an API  ArgumentTypeFuzzer::fuzzReturnType() to return a concrete
type that can bind to the formal return type of a function signature.

This diff fixes the issue https://github.com/facebookincubator/velox/issues/3533.

Reviewed By: bikramSingh91

Differential Revision: D42116543

fbshipit-source-id: 49a1c492dbb25c31ac3b9ceb2d2fb0087a032438