Introduce HashJoinNode::nullAware flag (#3580)
Summary:
It is necessary to distinguish between null aware (IN) and regular (EXISTS) semi joins.
To avoid adding 4 new values to the JoinType enum, we introduce nullAware flag
that applies to semi and anti joins and replace kNullAwareAntiJoin enum value
with a combination of kAntiJoin and nullAware = true.

See #3555 for more context.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3580

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D42200194

Pulled By: mbasmanova

fbshipit-source-id: c835dbfca9fbac974f08358abeddca026bf6cfcc