Fix in_map stream column id population (#3619)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3619

Populate the stream identifier metadata properly so feature reordering and feature stats aggregation can pick up in_map streams.

This currently causes the prod feature reordering to be slightly worse than the theoretical ceiling because it pushes the in_map stream of the reordered keys to the front of the stripe instead of grouping with the data streams in the back. However, since the reordered in_maps are still together in chunks, the gap should be very small.

Reviewed By: zzhao0

Differential Revision: D42177016

fbshipit-source-id: f325a43a3a25899a007a6c9ed20c3fe63b66246a