Fix handling of empty arrays and maps in subscript operator (#3644)
Summary:
A recent patch introduced a bug where the implementation of subscript
operator would check for an empty values vector in the input
array/map and directly return a null vector as a result. This meant
that checks for invalid indices were not performed which resulted in
divergence from the expected behavior for element_at or subscript
operators. This patch makes sure that the handling of such empty
arrays and maps is consistent with that of Presto, that is, checks
for invalid indices is done regardless of whether the array vector
has a non-empty values vector.

Fixes https://github.com/facebookincubator/velox/issues/3625

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3644

Test Plan: Added unit test and consolidated related ones in the same test case.

Reviewed By: kagamiori

Differential Revision: D42370492

Pulled By: bikramSingh91

fbshipit-source-id: e95095105104bb05b1d1452ffcc0cee77998e588