Fix the set of rows passed to context.saveAndReset in Expr::applyFunctionWithPeeling (#3708)
Summary:
Consider an expression like below:
 ```
 Expr:   try(is_null(cast(c0 as integer)))
 where c0 => [DICTIONARY VARCHAR: 10 elements, 3 nulls], [CONSTANT VARCHAR: 10 elements, abc]
 ```
Here `cast` will throw when attempting to cast a string to a number and is_null should return true for every null row _without_ an exception.

In this case then we expect that the expression should return a vector of 10 elements with 3 trues corresponding to the nulls. However in the common path we find that it returns an all null vector.

This is because , after peeling we only evaluate the constant vector once , which throws an exception, but when we save state in ScopedContextSaver, instead of only saving the rows on which we need to apply the function (i.e all non-null rows), we save _all_ rows. The error context then presumes that all rows threw an exception , which is wrong.
We fix this by ensuring that VectorSaver only saves the applied rows.
Fixes : https://github.com/facebookincubator/velox/issues/3682

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3708

Reviewed By: mbasmanova

Differential Revision: D42584087

Pulled By: kgpai

fbshipit-source-id: fa5b16f30178ba7e643f123dfe560b3a17bb4c6c