FIX: applyFunctionWithPeeling should flatten single arg functions input always. (#3734)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3734

if we have f1(f3(f2(c0,1)))
if c0 is constant encoding, then since all inputs are constant we peel all inputs initially into constant vectors.

then lets say f2 generates a dictionary vector.
right now f3 should peel the input and flatten it based on the vector function contract(single arg functions always receive flat), but we are not doing that
based on a heuristic decision that if the most recent peeling resulted in constant wrapping it's better not to peel again.
This diff changes the code to make an exception for that assumption when peeling is mandatory.

Reviewed By: Yuhta

Differential Revision: D42565341

fbshipit-source-id: e8843465c5162aeeafaf124294597fe5fc7fe926