Combine the root memoryPool for Velox and Koski. (#3754)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3754

Currently, in the veloski path, koski will create a MemoryManager in its QueryCtx. During the creation of the new MemoryManager, a new instance of MemoryPool would be created. Also, the MemoryPool created in Koski are reused in DwioConfig.

https://www.internalfb.com/code/fbsource/[06fd515cb6974e86da937b8afb9cb29d22bfa2ca]/fbcode/koski/connect/dw/dwio/DwioScan.cpp?lines=224

However, the memoryPool is not passed to the velox for reuse. Thus, Velox will use the default ManagerManager for its own, thus a new instance of MemoryPool would be created subsequently.

Abovementioned logic leads to the multiple destruction of MemoryManager as shown in this log P597271381.

This separation of the MemoryPool usage causes the lifecircle discrepency in the Veloski path, where the DwioScan relied on the MemoryPool koski created, whereas, the Velox are dependend on the Velox default MemoryPool.

Reviewed By: Yuhta

Differential Revision: D42590521

fbshipit-source-id: 662d2d0dd90ebd7f81852b675a76b2748543fbca