Avoid TaskRunner destruction on cancel
Drivers may leave the TaskRunner thread's interrupt flag set during the
course of processing, but doing so should not result in the TaskRunner
terminating its own processing loop until the TaskExecutor is closed.

Instead of allowing the interrupt to terminate the current task runner's
loop and re-creating a new runner to replace the interrupted one, we can
clear the interrupt flag after each iteration. Otherwise, TaskRunners
that were interrupted would have to replace themselves and end up
creating unnecessary threads in the cached threadpool executor in the
process.