Fix merge window optimizers
When the optimizer GatherAndMergeWindows merges or reorders two window
nodes, it will check if one window node takes the other window's output
as input or not. If yes, the two window nodes cannot be merged or reordered.
However, the current implementation to check this dependency misses the
input used in Frame. This commit fixes it by adding the Frame into account.