Introduce adaptive query executor
Adaptive query executor is part of the adaptive query execution framework.
It drives execution of a query in stages. At each step, it uses the
iterative plan fragmenter to get the next ready fragments for execution.
Each time a fragment finishes execution, it attempts to re-optimize the
plan before continuing execution.
There is a known issue with some complex joins that include nested queries,
which is being tracked in issue #18969.