Do not allow createType() for SHORT_ and LONG_DECIMAL (#3847)
Summary:
createType(SHORT_DECIMAL) used to return an instance of
ScalarType<SHORT_DECIMAL> which is not a correct representation of the decimal
type. It should be DecimalType<SHORT_DECIMAL>. After this change createType
will throw if input is SHORT_ or LONG_DECIMAL.

Also, improve error messages when creating decimal types with incorrect precision
or scale.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3847

Reviewed By: Yuhta

Differential Revision: D42803085

Pulled By: mbasmanova

fbshipit-source-id: 2ed439c8b248ca59a4c5aa7071e7e601d9f9312f